+++ Sources/MapboxCoreNavigation/EHorizon/RoadObjectStore.swift
@@ -93,7 +93,9 @@ public final class RoadObjectStore {
     }
 
     deinit {
-        native.setObserverForOptions(nil)
+        if native.hasObservers() {
+            native.removeObserver(for: self)
+        }
     }
 
     var native: MapboxNavigationNative.RoadObjectsStore {
@@ -104,9 +106,11 @@ public final class RoadObjectStore {
     
     private func updateObserver() {
         if delegate != nil {
-            native.setObserverForOptions(self)
+            native.addObserver(for: self)
         } else {
-            native.setObserverForOptions(nil)
+            if native.hasObservers() {
+                native.removeObserver(for: self)
+            }
         }
     }
 }