+++ Sources/MapboxNavigation/NavigationMapView+BuildingHighlighting.swift
@@ -33,7 +33,7 @@ extension NavigationMapView {
                                                     completion: { [weak self] result in
                                                         guard let _ = self else { return }
                                                         if case .success(let queriedFeatures) = result {
-                                                            if let identifier = queriedFeatures.first?.feature.identifier as? Int64 {
+                                                            if let identifier = queriedFeatures.first?.feature?.featureIdentifier {
                                                                 foundBuildingIds.insert(identifier)
                                                             }
                                                             group.leave()
@@ -113,7 +113,7 @@ extension NavigationMapView {
                 }
             )
             
-            highlightedBuildingsLayer.fillExtrusionColor = .constant(.init(color: buildingHighlightColor))
+            highlightedBuildingsLayer.fillExtrusionColor = .constant(.init(buildingHighlightColor))
             highlightedBuildingsLayer.fillExtrusionHeightTransition = StyleTransition(duration: 0.8, delay: 0)
             highlightedBuildingsLayer.fillExtrusionOpacityTransition = StyleTransition(duration: 0.8, delay: 0)
             try mapView.mapboxMap.style.addLayer(highlightedBuildingsLayer)